Fix environment variable name of plugin daemon URL in worker (#93)
*Issue #, if available:*

**Description of changes:**
In the current implementation, the environment variables used to specify
the plugin daemon differ between the `api` and `worker` containers, even
though both use the same container image.

For the `api` container, `PLUGIN_DAEMON_URL` is used:

[https://github.com/aws-samples/dify-self-hosted-on-aws/blob/20ba89253de5191650cdc4876baf80c7048d2f0f/lib/constructs/dify-services/api.ts#L125](https://github.com/aws-samples/dify-self-hosted-on-aws/blob/20ba89253de5191650cdc4876baf80c7048d2f0f/lib/constructs/dify-services/api.ts#L125)

For the `worker` container, `PLUGIN_API_URL` is used:

[https://github.com/aws-samples/dify-self-hosted-on-aws/blob/20ba89253de5191650cdc4876baf80c7048d2f0f/lib/constructs/dify-services/api.ts#L224](https://github.com/aws-samples/dify-self-hosted-on-aws/blob/20ba89253de5191650cdc4876baf80c7048d2f0f/lib/constructs/dify-services/api.ts#L224)

According to the Dify `api` implementation, the correct environment
variable should be `PLUGIN_DEAMON_URL`:

[https://github.com/langgenius/dify/blob/4a89403566c85eff76a959bc62be76f1e8af5ce6/api/configs/feature/__init__.py#L209-L212](https://github.com/langgenius/dify/blob/4a89403566c85eff76a959bc62be76f1e8af5ce6/api/configs/feature/__init__.py#L209-L212)

If `PLUGIN_DAEMON_URL` is not properly specified, the default value
`http://localhost:5002` is used. This works under the current
architecture, where the `worker` and plugin daemon run in the same ECS
task.
However, it may be misleading if someone attempts to run the `worker`
and plugin daemon in separate ECS tasks.

This PR fixes the environment variable in `worker` to the correct one.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.